chore: rename some story component name
