+++ modules/gallery/src/commonMain/kotlin/story/StoryList.kt
@@ -14,13 +14,12 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import org.jetbrains.compose.storytale.Story
-import org.jetbrains.compose.storytale.gallery.StoryComponentListItem
 import org.jetbrains.compose.storytale.gallery.ui.component.CenterRow
 import org.jetbrains.compose.storytale.gallery.ui.component.Gap
 import org.jetbrains.compose.storytale.gallery.ui.component.NumberChip
 
 @Composable
-fun StoryGalleryList(
+fun StoryList(
   activeStoryIndex: Int,
   stories: List<Story>,
   onSelectStory: (index: Int) -> Unit
@@ -44,7 +43,7 @@ fun StoryGalleryList(
         verticalArrangement = Arrangement.spacedBy(4.dp)
       ) {
         stories.forEachIndexed { index, s ->
-          StoryComponentListItem(
+          StoryListItem(
             story = s,
             selected = index == activeStoryIndex,
             modifier = Modifier.fillMaxWidth().padding(horizontal = 4.dp),
